semantic-seg/requirements.txt
absl-py==0.5.0
astor==0.7.1
astroid==2.0.4
certifi==2018.10.15
cycler==0.10.0
gast==0.2.0
grpcio==1.12.1
h5py==2.8.0
imageio==2.4.1
isort==4.3.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
matplotlib==3.0.0
mccabe==0.6.1
mkl-fft==1.0.6
mkl-random==1.0.1
numpy==1.15.2
Pillow==5.3.0
protobuf==3.6.0
pylint==2.1.1
pyparsing==2.2.2
python-dateutil==2.7.3
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
typed-ast==1.1.0
UnrealCV==0.3.10
Werkzeug==0.14.1
wrapt==1.10.11
